fix(ci): satisfy changed Go lint
